backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
Automat
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
coloredlogs
constantly
crochet
cryptography
cssselect
dataclasses-json
deepdiff
defusedxml
distro
Django
django-cors-headers
djangorestframework
durationpy
faiss-cpu
filelock
flatbuffers
frozenlist
fsspec
google-auth
googleapis-common-protos
greenlet
grpcio
h11
hf-xet
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
hyperlink
idna
importlib_metadata
importlib_resources
Incremental
itemadapter
itemloaders
Jinja2
jmespath
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-chroma
langchain-classic
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
mpmath
multidict
mypy_extensions
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvshmem-cu12
nvidia-nvtx-cu12
oauthlib
ollama
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orderly-set
orjson
ormsgpack
overrides
packaging
parsel
posthog
propcache
Protego
protobuf
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
PyDispatcher
Pygments
pyOpenSSL
PyPika
pyproject_hooks
python-dateutil
python-docx
python-dotenv
PyYAML
queuelib
referencing
regex
requests
requests-file
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
Scrapy
sentence-transformers
service-identity
setuptools
shellingham
six
soupsieve
SQLAlchemy
sqlparse
sympy
tenacity
threadpoolctl
tldextract
tokenizers
torch
tqdm
transformers
triton
Twisted
typer
typing-inspect
typing-inspection
typing_extensions
urllib3
uuid_utils
uvicorn
uvloop
w3lib
watchfiles
websocket-client
websockets
wrapt
xxhash
yarl
zipp
zope.interface
zstandard
psycopg2-binary
gunicorn